Setup server to respond to non api web requests
