+++ http/middleware.go
@@ -1,0 +1,30 @@
+package http
+
+import (
+	gin "github.com/gin-gonic/gin"
+	"os"
+	"path/filepath"
+)
+
+/**
+ * If a route doesn't match the /api pattern and isn't a static
+ * asset I want to send to the index.html page and let the client
+ * side handle it without the api server returning a 404.
+ *
+ * This middleware looks for those http 404s and serves the index
+ */
+func Intercept404AndServeReact(c *gin.Context) {
+	// Ignoring this error because we would have seen it already
+	ex, _ := os.Executable()
+
+	// If we have a 404 then serve the index page and
+	// let client side routing take over
+	if c.Writer.Status() == 404 {
+		fname := filepath.Join(
+			filepath.Dir(ex),
+			"test/dist/index.html")
+
+		c.File(fname)
+		c.Abort()
+	}
+}
