+++ internal/migration.go
@@ -1,0 +1,34 @@
+package internal
+
+/**
+ * Schema used to migrate the db
+ */
+var Schema string = `
+CREATE TABLE IF NOT EXISTS kids (
+	id INT GENERATED BY DEFAULT AS IDENTITY,
+	name VARCHAR NOT NULL,
+	PRIMARY KEY(id)
+);
+
+CREATE TABLE IF NOT EXISTS pokemon (
+	id INT GENERATED BY DEFAULT AS IDENTITY,
+	name VARCHAR NOT NULL,
+	PRIMARY KEY(id)
+);
+
+CREATE TABLE IF NOT EXISTS cards (
+	id INT GENERATED BY DEFAULT AS IDENTITY,
+	tcg_id VARCHAR NOT NULL,
+	pokemon_id INT NOT NULL,
+	kind VARCHAR NOT NULL,
+	PRIMARY KEY(id),
+	CONSTRAINT fk_pokemon FOREIGN KEY(pokemon_id) REFERENCES pokemon(id)
+);
+
+CREATE TABLE IF NOT EXISTS cards_kids (
+	card_id INT NOT NULL,
+	kid_id INT NOT NULL,
+	CONSTRAINT fk_card FOREIGN KEY(card_id) REFERENCES cards(id),
+	CONSTRAINT fk_kid FOREIGN KEY(kid_id) REFERENCES kids(id)
+)
+`
