+++ pokemon-web.go
@@ -1,53 +1,52 @@
 package main
 
 import (
+	"log"
+	"os"
+	"path/filepath"
+
+	"github.com/ghargrove/pokemon-web/http"
 	"github.com/ghargrove/pokemon-web/http/rest"
 	"github.com/ghargrove/pokemon-web/internal"
 	"github.com/gin-gonic/gin"
 	"github.com/joho/godotenv"
-	"log"
 )
-
-var schema string = `
-CREATE TABLE IF NOT EXISTS kids (
-	id INT GENERATED BY DEFAULT AS IDENTITY,
-	name VARCHAR NOT NULL,
-	PRIMARY KEY(id)
-);
-
-CREATE TABLE IF NOT EXISTS pokemon (
-	id INT GENERATED BY DEFAULT AS IDENTITY,
-	name VARCHAR NOT NULL,
-	PRIMARY KEY(id)
-);
-
-CREATE TABLE IF NOT EXISTS cards (
-	id INT GENERATED BY DEFAULT AS IDENTITY,
-	tcg_id VARCHAR NOT NULL,
-	pokemon_id INT NOT NULL,
-	kind VARCHAR NOT NULL,
-	PRIMARY KEY(id),
-	CONSTRAINT fk_pokemon FOREIGN KEY(pokemon_id) REFERENCES pokemon(id)
-);
-
-CREATE TABLE IF NOT EXISTS cards_kids (
-	card_id INT NOT NULL,
-	kid_id INT NOT NULL,
-	CONSTRAINT fk_card FOREIGN KEY(card_id) REFERENCES cards(id),
-	CONSTRAINT fk_kid FOREIGN KEY(kid_id) REFERENCES kids(id)
-)
-`
 
 func main() {
-	err := godotenv.Load()
+	// Usefule for getting the exec path
+	// https://stackoverflow.com/questions/18537257/how-to-get-the-directory-of-the-currently-running-file
+	ex, err := os.Executable()
+	if err != nil {
+		log.Fatal("Could not read the executable")
+	}
+	// The directory we can use to absolutely reference app files
+	exPath := filepath.Dir(ex)
+
+	err = godotenv.Load(filepath.Join(exPath, ".env"))
 	if err != nil {
 		log.Fatal("Error loading env file")
 	}
 
 	db := internal.CreateDB()
-	db.MustExec(schema)
+	db.MustExec(internal.Schema)
 
 	r := gin.Default()
+
+	// Statically serve the compiled assets
+	r.Static("/assets", filepath.Join(exPath, "test/dist/assets"))
+
+	r.Use(http.Intercept404AndServeReact)
+
+	r.Use(func(c *gin.Context) {
+		// If we have a 404 then serve the index page and
+		// let client side routing take over
+		if c.Writer.Status() == 404 {
+			fname := filepath.Join(exPath, "test/dist/index.html")
+
+			c.File(fname)
+			c.Abort()
+		}
+	})
 
 	r.GET("/api/sets", rest.HandleAllSets)
 	r.GET("/api/sets/:id", rest.HandleSingleSet)
@@ -61,5 +60,6 @@ func main() {
 	r.POST("/api/kids/:id/cards", func(c *gin.Context) {
 		rest.CreateKidCard(c, db)
 	})
+
 	r.Run(":8082")
 }
